+++ CHANGELOG.md
@@ -14,11 +14,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the ability to build multiple signatures from a data source on the "ConstructorSchemes"
 - Multiselect: added setting "variant search"
 - Multiselect: if after entering the search in the multiline there is one option, then select it by entry
+- Added tree icon in header
+- Added auto-centered Map
+- Added centered Map by token
+- Added search by objects Map
 
 ### Changed
 
 ### Fixed
 - Heatmap: events don't pass token values
+- Map: fixed incorrect update token with action "click"
+- RiskReview: Fixed incorrect resizing when switching between "full screen", "modal" and "default mode"
 
 ## [1.129.0]